resolve: implement leaderboard component and enhance sidebar navigation structure
