+++ components/leaderboard/leaderboard-item.tsx
